Add tx_halt and tx_resume flow control features
This allows the board, or a message being sent to the board, to halt transmission (tx_halt) of further data to the board. Only the board can trigger tx_resume by sending the appropriate signal. This is for future functions that might disable interrupts long enough to cause data loss.